Move constants file to data folder
