+++ data_models/__init__.py
@@ -1,7 +1,7 @@
 from .action import Action, Add, Remove, Confirm
 from .item import Item
 from .order import Order, OrderStatus
-from .Constants import Constants, ConstantUnits
+from data.constants import Constants, ConstantUnits
 
 __all__ = [
     "Action",
